Update String and Range object types.

+++ include/utils.h
@@ -5,4 +5,4 @@
 
 bool ends_with(std::string_view str, std::string_view suffix);
 bool starts_with(std::string_view str, std::string_view prefix);
-std::array<i32, 3> constructRange(std::string_view tokText);
+std::array<i64, 3> constructRange(std::string_view tokText);
